fix CodeStyle
