+++ src/httpserver/handlers/EpicsHandler.java
@@ -1,14 +1,15 @@
-package httpServer.handlers;
+package httpserver.handlers;
 
-import classes.tasks.Task;
+import classes.tasks.Epic;
 import com.sun.net.httpserver.HttpExchange;
 import managers.TaskManager;
+
 import java.io.IOException;
 import java.util.Arrays;
 
-public class TasksHandler extends BasicHandler {
+public class EpicsHandler extends BasicHandler {
 
-    public TasksHandler(TaskManager taskManager) {
+    public EpicsHandler(TaskManager taskManager) {
         super(taskManager);
     }
 
@@ -16,9 +17,9 @@ public class TasksHandler extends BasicHandler {
     protected void getById(HttpExchange exchange, Integer id) throws IOException {
         String response;
         try {
-            Task task = taskManager.getTaskById(id);
-            if (task != null) {
-                response = gson.toJson(task);
+            Epic epic = taskManager.getEpicById(id);
+            if (epic != null && taskManager.getEpics().contains(epic)) {
+                response = gson.toJson(epic);
                 sendResponse(exchange, response);
             } else {
                 handleNotFound(exchange);
@@ -32,8 +33,8 @@ public class TasksHandler extends BasicHandler {
     protected void getList(HttpExchange exchange) throws IOException {
         String response;
         try {
-            if (!taskManager.getTasks().isEmpty()) {
-                response = gson.toJson(taskManager.getTasks());
+            if (!taskManager.getEpics().isEmpty()) {
+                response = gson.toJson(taskManager.getEpics());
                 sendResponse(exchange, response);
             } else {
                 handleNotFound(exchange);
@@ -46,9 +47,9 @@ public class TasksHandler extends BasicHandler {
     @Override
     protected void postUpdate(HttpExchange exchange, Integer id, String body) throws IOException {
         try {
-            Task task = gson.fromJson(body, Task.class);
-            taskManager.updateTask(task);
-            String response = gson.toJson(taskManager.getTaskById(id));
+            Epic epic = gson.fromJson(body, Epic.class);
+            taskManager.updateEpic(epic);
+            String response = gson.toJson(taskManager.getEpicById(id));
             sendResponse(exchange, response);
         } catch (Exception e) {
             handleNotFound(exchange);
@@ -58,10 +59,10 @@ public class TasksHandler extends BasicHandler {
     @Override
     protected void postCreate(HttpExchange exchange, String body) throws IOException {
         try {
-            Task task = gson.fromJson(body, Task.class);
-            taskManager.addTask(task);
+            Epic epic = gson.fromJson(body, Epic.class);
+            taskManager.addEpic(epic);
 
-            String response = gson.toJson(task);
+            String response = gson.toJson(epic);
             sendResponse(exchange, response);
         } catch (Exception e) {
             handleNotFound(exchange);
@@ -81,10 +82,10 @@ public class TasksHandler extends BasicHandler {
                 Integer id = Integer.parseInt(elements[1]);
                 String response;
 
-                Task task = taskManager.getTaskById(id);
-                if (task != null && taskManager.getTasks().contains(task)) {
-                    response = gson.toJson(task);
-                    taskManager.removeTaskById(id);
+                Epic epic = taskManager.getEpicById(id);
+                if (epic != null && taskManager.getEpics().contains(epic)) {
+                    response = gson.toJson(epic);
+                    taskManager.removeEpicById(id);
                     sendResponse(exchange, response);
                 } else {
                     handleNotFound(exchange);
@@ -96,6 +97,18 @@ public class TasksHandler extends BasicHandler {
             handleBadRequest(exchange);
         } catch (Exception e) {
             handleInternalServerError(exchange);
+        }
+    }
+
+    @Override
+    protected void getEpicSubtasks(HttpExchange exchange, Integer id) throws IOException {
+        String response;
+        if (taskManager.getEpics().contains(taskManager.getEpicById(id))
+                && !taskManager.getSubtasksByEpic(taskManager.getEpicById(id)).isEmpty()) {
+            response = gson.toJson(taskManager.getSubtasksByEpic(taskManager.getEpicById(id)));
+            sendResponse(exchange, response);
+        } else {
+            handleNotFound(exchange);
         }
     }
 }
