+++ src/httpserver/handlers/SubtasksHandler.java
@@ -1,15 +1,14 @@
-package httpServer.handlers;
+package httpserver.handlers;
 
-import classes.tasks.Epic;
+import classes.tasks.Subtask;
 import com.sun.net.httpserver.HttpExchange;
 import managers.TaskManager;
-
 import java.io.IOException;
 import java.util.Arrays;
 
-public class EpicsHandler extends BasicHandler {
+public class SubtasksHandler extends BasicHandler {
 
-    public EpicsHandler(TaskManager taskManager) {
+    public SubtasksHandler(TaskManager taskManager) {
         super(taskManager);
     }
 
@@ -17,9 +16,9 @@ public class EpicsHandler extends BasicHandler {
     protected void getById(HttpExchange exchange, Integer id) throws IOException {
         String response;
         try {
-            Epic epic = taskManager.getEpicById(id);
-            if (epic != null && taskManager.getEpics().contains(epic)) {
-                response = gson.toJson(epic);
+            Subtask subtask = taskManager.getSubtaskById(id);
+            if (subtask != null && taskManager.getSubtasks().contains(subtask)) {
+                response = gson.toJson(subtask);
                 sendResponse(exchange, response);
             } else {
                 handleNotFound(exchange);
@@ -33,8 +32,8 @@ public class EpicsHandler extends BasicHandler {
     protected void getList(HttpExchange exchange) throws IOException {
         String response;
         try {
-            if (!taskManager.getEpics().isEmpty()) {
-                response = gson.toJson(taskManager.getEpics());
+            if (!taskManager.getSubtasks().isEmpty()) {
+                response = gson.toJson(taskManager.getSubtasks());
                 sendResponse(exchange, response);
             } else {
                 handleNotFound(exchange);
@@ -47,9 +46,10 @@ public class EpicsHandler extends BasicHandler {
     @Override
     protected void postUpdate(HttpExchange exchange, Integer id, String body) throws IOException {
         try {
-            Epic epic = gson.fromJson(body, Epic.class);
-            taskManager.updateEpic(epic);
-            String response = gson.toJson(taskManager.getEpicById(id));
+            Subtask subtask = gson.fromJson(body, Subtask.class);
+            taskManager.updateSubtask(subtask);
+
+            String response = gson.toJson(taskManager.getSubtaskById(id));
             sendResponse(exchange, response);
         } catch (Exception e) {
             handleNotFound(exchange);
@@ -59,10 +59,10 @@ public class EpicsHandler extends BasicHandler {
     @Override
     protected void postCreate(HttpExchange exchange, String body) throws IOException {
         try {
-            Epic epic = gson.fromJson(body, Epic.class);
-            taskManager.addEpic(epic);
+            Subtask subtask = gson.fromJson(body, Subtask.class);
+            taskManager.addSubtask(subtask);
 
-            String response = gson.toJson(epic);
+            String response = gson.toJson(subtask);
             sendResponse(exchange, response);
         } catch (Exception e) {
             handleNotFound(exchange);
@@ -82,10 +82,10 @@ public class EpicsHandler extends BasicHandler {
                 Integer id = Integer.parseInt(elements[1]);
                 String response;
 
-                Epic epic = taskManager.getEpicById(id);
-                if (epic != null && taskManager.getEpics().contains(epic)) {
-                    response = gson.toJson(epic);
-                    taskManager.removeEpicById(id);
+                Subtask subtask = taskManager.getSubtaskById(id);
+                if (subtask != null && taskManager.getSubtasks().contains(subtask)) {
+                    response = gson.toJson(subtask);
+                    taskManager.removeSubtaskById(id);
                     sendResponse(exchange, response);
                 } else {
                     handleNotFound(exchange);
@@ -97,18 +97,6 @@ public class EpicsHandler extends BasicHandler {
             handleBadRequest(exchange);
         } catch (Exception e) {
             handleInternalServerError(exchange);
-        }
-    }
-
-    @Override
-    protected void getEpicSubtasks(HttpExchange exchange, Integer id) throws IOException {
-        String response;
-        if (taskManager.getEpics().contains(taskManager.getEpicById(id))
-                && !taskManager.getSubtasksByEpic(taskManager.getEpicById(id)).isEmpty()) {
-            response = gson.toJson(taskManager.getSubtasksByEpic(taskManager.getEpicById(id)));
-            sendResponse(exchange, response);
-        } else {
-            handleNotFound(exchange);
         }
     }
 }
